requirements.txt
Flask==2.3.2
psutil==5.9.8
psycopg2-binary==2.9.6